+++ src/checksum.h
@@ -3,8 +3,8 @@
 
 #include <stdint.h>
 
-uint16_t get_checksum16(uint16_t* arr, int size);
-uint16_t get_crc16(uint8_t* arr, int size);
+uint16_t get_checksum16(uint16_t* arr, size_t size);
+uint16_t get_crc16(uint8_t* arr, size_t size);
 
 #endif
